More math added.  Needs two more chunks.
I’m not sure if that last integral in the math section can actually be
evaluated (if p(u) is uniform).  I haven’t found any identities for it
yet.

The remaining stuff that needs to go in the math section is (a) the
spatial average calculation, and (b) the explanation of how they are
different.