Minor fix to montecarlo
This doesn't work as easily as I was hoping. I need to take an average
of each of the value positions within the matrix H_hat and determine how
far off they are from 5%. Which is obvious in hindsite that it wouldn't
be so easy to just copy KC's code. I'll look at it again tomorrow.