Fixed, and more or less done
averaging variable is now used in the testing runs. default scenario may
need further tweaking but it looks like it works well now. I'll try to
find a default that is decent but not perfect so as to give a good
representation of what does and does not have a strong effect on the
sim.